All encoders implemented, read, displayed, pushing over serial
I have run into the predictable issue that the encoders can just twist off into infinity and there's no way to clamp them. I think I will have to change an internal value by the difference between the current and last poll of the encoder so as to effectively clamp the value. This will also allow me to easily modify the rotation multiplier.