Encoders respect 0-100 bound, muting fixed, layer sel by encoder
The RotaryEncoder class has been replaced with VolKnob which handles the important math to determine volume and mute state itself. It also has another method that allows it to be read as an offset, which is now used to navigate the layer selection using the master volume knob. Muting is now visually displayed by graying out the relevant bar.